fix: remove transparent border around resized images

+++ config/rust/src/lib.rs
@@ -15,7 +15,7 @@ pub const COOKIE_DOMAIN: &str = "jicloud.org";
 pub const CORS_ORIGINS: [&str; 2] = ["https://jicloud.org", "https://sandbox.jicloud.org"];
 pub const DB_POOL_CONNECTIONS: u32 = 5;
 
-pub const IMAGE_BODY_SIZE_LIMIT: u64 = 1024 * 1024 * 10; // 10 MB
+pub const IMAGE_BODY_SIZE_LIMIT: usize = 1024 * 1024 * 10; // 10 MB
 
 pub const REMOTE_DB_USER: &str = "postgres";
 pub const REMOTE_DB_NAME: &str = "jicloud";
